Remove unneeded function for now
